+++ src/main/java/blackjack/Application.java
@@ -2,12 +2,12 @@ package blackjack;
 
 import blackjack.controller.BlackjackController;
 import blackjack.manager.BlackJackInitManager;
-import blackjack.manager.CardsGenerator;
+import blackjack.manager.SingDeckGenerator;
 
 public class Application {
 
     public static void main(String[] args) {
-        BlackJackInitManager blackJackInitManager = new BlackJackInitManager(new CardsGenerator());
+        BlackJackInitManager blackJackInitManager = new BlackJackInitManager(new SingDeckGenerator());
         BlackjackController blackjackController = new BlackjackController(blackJackInitManager);
 
         blackjackController.run();
