+++ src/main/java/blackjack/domain/Deck.java
@@ -3,28 +3,15 @@ package blackjack.domain;
 import blackjack.common.ErrorMessage;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
 import java.util.stream.IntStream;
 
 public class Deck {
 
-    private static final int DECK_SIZE = 52;
     private final List<Card> cards;
 
     public Deck(List<Card> cards) {
-        validate(cards);
         this.cards = new ArrayList<>(shuffled(cards));
-    }
-
-    private void validate(List<Card> cards) {
-        if (cards.size() != DECK_SIZE) {
-            throw new IllegalArgumentException(ErrorMessage.INVALID_DECK_SIZE.getMessage());
-        }
-
-        if (new HashSet<>(cards).size() != DECK_SIZE) {
-            throw new IllegalArgumentException(ErrorMessage.DUPLICATED_CARD_EXISTED.getMessage());
-        }
     }
 
     private List<Card> shuffled(List<Card> cards) {
