+++ src/main/java/blackjack/manager/DeckGenerator.java
@@ -1,0 +1,9 @@
+package blackjack.manager;
+
+import blackjack.domain.Card;
+import java.util.List;
+
+public interface DeckGenerator {
+
+    List<Card> generate();
+}
